requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
werkzeug==2.3.7
python-multipart==0.0.6
requests==2.31.0
Pillow==10.0.1
PyPDF2==3.0.1
python-docx==0.8.11
supabase==1.0.4
python-dotenv==1.0.0
PyJWT==2.8.0
gunicorn==21.2.0
psutil==5.9.6
pdf2docx==0.5.6
PyMuPDF==1.23.8
pytesseract==0.3.10
pandas==2.1.1
python-pptx==0.6.21
img2pdf==0.4.4
# Enhanced conversion dependencies
pdf2image==1.16.3
openpyxl==3.1.2
xlrd==2.0.1
xlsxwriter==3.1.9
reportlab==4.0.4
weasyprint==60.0
beautifulsoup4==4.12.2
lxml==4.9.3
# Additional dependencies for production
docx2pdf==0.1.8
schedule==1.2.0
python-magic==0.4.27
# Production server dependencies
gevent==23.7.0
eventlet==0.33.3